requirements.txt
# dev
mesos.interface>=1.3.0
# testing
mock>=2.0.0